requirements.txt
django-braces==1.10.0
django-environ==0.4.1
django-extensions==1.7.5
django-model-utils==2.6
Django==1.10.4
ipdb==0.10.1
ipython==5.1.0
psycopg2==2.6.2
